API template format

API template format

baseurl

required: TRUE

type: character string

description: the base url should be the base url, on to which any routes are added (e.g., base url of 'http://stuff.com', and the route 'things' would make a url of 'http://stuff.com/things')

data

required: TRUE

type: character string

description: one of json or xml

routes

required: TRUE

type: array

description: any number of routes, see route below for description

a single route

required: TRUE

type: array

description: Follows the following pattern: name: path: id: options: NULL class: character required: FALSE params: query: options: NULL class: character required: FALSE